+++ src/methods/zero_shot/inductive_clip.py
@@ -118,7 +118,7 @@ class CLIP(BASE):
             self.u = deepcopy(query)
         else:
             self.u = torch.zeros((n_task, query.shape[1], n_ways)).to(self.device)
-            text_features = clip_weights(self.model, self.args.classnames, self.args.template, self.device).double()
+            text_features = clip_weights(self.model, self.args.classnames, self.args.template, self.device)
             for task in range(n_task):
                 image_features = query[task] / query[task].norm(dim=-1, keepdim=True)
                 sim = (self.args.T * (image_features @ text_features.T)).softmax(dim=-1) # N* K
